Feb 23 - 5:35 PM

+++ lib/screens/Home_Screen/Users/Add Users/add_users_form.dart
@@ -20,6 +20,7 @@ class AddUsersForm extends StatefulWidget {
 class _AddUsersFormState extends State<AddUsersForm> {
   final _formKey = GlobalKey<FormState>();
   final List<String> errors = [];
+  FirebaseApp fbApp = Firebase.app('Secondary');
 
   String email;
   String password;
@@ -184,22 +185,37 @@ class _AddUsersFormState extends State<AddUsersForm> {
   }
 
   Future createUser(email, password, context) async {
-    FirebaseApp app = await Firebase.initializeApp(
-        name: 'Secondary', options: Firebase.app().options);
-    try {
+    if (Firebase.apps.length == 1) {
+      FirebaseApp app = await Firebase.initializeApp(
+          name: 'Secondary', options: Firebase.app().options);
+
       await FirebaseAuth.instanceFor(app: app)
           .createUserWithEmailAndPassword(email: email, password: password)
-          .then((value) {
+          .then((value) async {
         SetData().addStudent(context,
             email: email,
             batch: batch,
             department: department,
             regNo: password);
+        await app.delete();
+      }).catchError((e) {
+        FirebaseAuth.instanceFor(app: app).signOut();
+        Navigator.pop(context);
+        Snack_Bar.show(context, e.message);
       });
-    } catch (e) {
-      Snack_Bar.show(context, e.message);
     }
-    await app.delete();
+
+    await FirebaseAuth.instanceFor(app: fbApp)
+        .createUserWithEmailAndPassword(email: email, password: password)
+        .then((value) async {
+      SetData().addStudent(context,
+          email: email, batch: batch, department: department, regNo: password);
+      await fbApp.delete();
+    }).catchError((e) {
+      FirebaseAuth.instanceFor(app: fbApp).signOut();
+      Navigator.pop(context);
+      Snack_Bar.show(context, e.message);
+    });
   }
 
   show() {
